# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/category-priorities/{categoryName}": {
      "post": {
        "operationId": "CategoryPrioritiesController_addPriority",
        "parameters": [
          {
            "name": "categoryName",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryPriorityPayload"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Priority added successfully." },
          "400": { "description": "Malformed payload." },
          "401": { "description": "User is not authenticated." },
          "404": { "description": "Category name not found." },
          "409": {
            "description": "Category priority already exists for this user."
          }
        },
        "tags": ["categories"]
      },
      "delete": {
        "operationId": "CategoryPrioritiesController_deletePriority",
        "parameters": [
          {
            "name": "categoryName",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Priority deleted successfully." },
          "401": { "description": "User is not authenticated." },
          "404": { "description": "Category name not found." }
        },
        "tags": ["categories"]
      }
    },
    "/v1/category-priorities/export": {
      "get": {
        "operationId": "CategoryPrioritiesController_exportCategoryPrioritiesMatrix",
        "parameters": [],
        "responses": {
          "200": { "description": "Matrix exported successfully." },
          "401": { "description": "User is not authenticated." },
          "403": { "description": "User is not administrator." }
        },
        "tags": ["categories"]
      }
    },
...
  },
  "info": {
    "title": "Brief me API",
    "description": "API documentation for the Brief me app.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CategoryPriorityPayload": {
        "type": "object",
        "properties": {
          "priority": { "type": "number", "minimum": 0, "maximum": 10 }
        },
        "required": ["priority"]
      }
    }
  }
}